chore(warn): Fix MSVC warning C4838.
Macro needed a cast adding to avoid incorrect promotion from bool to int rather than UINT before bit shifts. result should be the same minus the warnings.

+++ Code/Tools/LevelEdit/PresetsLibForm.cpp
@@ -108,7 +108,7 @@ private:
 #ifndef TreeView_SetOverlay
    #define TreeView_SetOverlay(hwndLV, i, overlay) \
       {	\
-			TVITEMEX item_info = { TVIF_STATE, i, INDEXTOOVERLAYMASK(overlay), TVIS_OVERLAYMASK, 0 };	\
+			TVITEMEX item_info = { TVIF_STATE, i, INDEXTOOVERLAYMASK(static_cast<UINT>(overlay)), TVIS_OVERLAYMASK, 0 };	\
 			TreeView_SetItem(hwndLV, &item_info);	\
 		}
 #endif
